pyproject.toml
[tool.poetry]
name = "mcstatus_bot"
version = "0.1.0"
description = ""
authors = ["Egorrko"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = {version = "^3.0.0b7", allow-prereleases = true}
"discord.py" = "^2.2.3"
mcstatus = "^10.0.3"
pydantic = "^1.10.5"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
